// src/sync/redis/interface.ts
// Generated by https://quicktype.io

export interface IUpbitWsCryptoResponse {
  code: string; // symbol
  opening_price: number; // open
  high_price: number; // dayHigh
  low_price: number; // dayLow
  trade_price: number; // price
  prev_closing_price: number; //previousClose
  signed_change_price: number; //change
  signed_change_rate: number; //changesPercentage
  timestamp: number; //timestamp
  acc_trade_price: number;
  change: string;
  change_price: number;
  change_rate: number;
  ask_bid: string;
  trade_volume: number;
  acc_trade_volume: number;
  trade_date: string;
  trade_time: string;
  trade_timestamp: number;
  acc_ask_volume: number;
  acc_bid_volume: number;
  highest_52_week_price: number;
  highest_52_week_date: string;
  lowest_52_week_price: number;
  lowest_52_week_date: string;
  trade_status: string;
  market_state: string;
  market_state_for_ios: string;
  is_trading_suspended: boolean;
  delisting_date: null;
  market_warning: string;
  acc_trade_price_24h: number;
  acc_trade_volume_24h: number;
}

// export interface Quote {
//     symbol: string;
//     name: string;
//     price: number;
//     changesPercentage: number;
//     change: number;
//     dayLow: number;
//     dayHigh: number;
//     yearHigh: number;
//     yearLow: number;
//     marketCap: number;
//     priceAvg50: number;
//     priceAvg200: number;
//     volume: number;
//     avgVolume: number;
//     exchange: string;
//     open: number;
//     previousClose: number;
//     eps: number;
//     pe: number;
//     earningsAnnouncement: string;
//     sharesOutstanding: number;
//     timestamp: number;
//   }